requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
chromadb>=0.5.0
sentence-transformers>=2.6.1
PyPDF2>=3.0.0
python-docx>=1.1.0
networkx>=3.2.1
pyvis>=0.3.2
python-dotenv>=1.0.1
pandas>=2.2.0
plotly>=5.19.0
